pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fenics-ffcx"
version = "0.8.0.dev0"
description = "The FEniCSx Form Compiler"
readme = "README.md"
requires-python = ">=3.8.0"
license = {file = "LICENSE"}
authors = [{email="fenics-dev@googlegroups.com"}, {name="FEniCS Project"}]
dependencies = [
    "numpy>=1.21",
    "cffi",
    "fenics-basix >= 0.8.0.dev0, <0.9.0",
    "fenics-ufl >= 2023.3.0.dev0, <2023.4.0",
]

[project.urls]
homepage = "https://fenicsproject.org"
repository = "https://github.com/fenics/ffcx.git"
documentation = "https://docs.fenicsproject.org"

[project.scripts]
ffcx = "ffcx:__main__.main"

[project.optional-dependencies]
lint = ["flake8", "pydocstyle[toml]"]
docs = ["sphinx", "sphinx_rtd_theme"]
optional = ["numba", "pygraphviz==1.7"]
test = ["pytest >= 6.0", "sympy"]
ci = [
    "coveralls",
    "coverage",
    "isort",
    "pytest-cov",
    "pytest-xdist",
    "types-setuptools",
    "mypy",
    "fenics-ffcx[docs]",
    "fenics-ffcx[optional]",
    "fenics-ffcx[lint]",
    "fenics-ffcx[test]",
]

[tool.setuptools]
packages = [
    "ffcx",
    "ffcx.codegeneration",
    "ffcx.codegeneration.C",
    "ffcx.ir",
    "ffcx.ir.analysis",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = [
    "test"
]
norecursedirs = [
    "libs",
    "docs"
]
log_cli = true

[tool.pydocstyle]
convention = "numpy"
# TODO: Work on removing these ignores
add-ignore = [
    "D100",  # Missing docstrings in modules
    "D101",  # Missing docstrings in classes
    "D102",  # Missing docstrings in methods
    "D103",  # Missing docstrings in functions
    "D104",  # Missing docstrings in packages
    "D105",  # Missing docstrings in magic methods
    "D401",  # Google convention
    "D406",  # Google convention
    "D407"   # Google convention
]

[tool.mypy]
# Suggested at https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
# Goal would be to make all of the below True long-term
disallow_untyped_defs = false
disallow_any_unimported = false
no_implicit_optional = false
check_untyped_defs = false
warn_return_any = false
warn_unused_ignores = false
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    'basix',
    'cffi',
    'numba.*',
    'pygraphviz',
    'ufl.*'
]
ignore_missing_imports = true
